requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
reportlab>=4.0.0
python-dotenv>=1.0.0
lxml>=4.9.0
selenium>=4.40.0
webdriver-manager>=4.0.0
fpdf2>=2.7.0
pypdf>=3.0.0
fonttools>=4.0.0
Pillow>=10.0.0
